Move definitions for Evaluation to Manipulation
